Make cache specific to a src/dest pair.
The cache will need to be low-level if we use openmp and handling
multiple sources in a single cache makes it very complex.  We instead
instantiate many cache pairs and use them as needed.  Also added a
`set_view_raw` to CArrays.  This allows us to point the data to a raw
memory location from Cython.