Cleanup API and make neighbor computation safer.
Instead of using raw pointers we now use CArrays for the neighbors.
This is achieved by using void* pointers to point to the neighbor arrays
for the thread local storage.  This requires explicit casting but makes
it much easier to write the code.  The API is also nicer and one does
not have to make assumptions on the number of neighbors for safe
operation.  Also using the alignment functions from carray in other
places as well.