WIP: first cut of support for properties with strides.
The idea is if you want a 3-element property per particle, you currently
need to add this in as a constant which is a pain.  With this, you can
simply do `pa.add_property('A', data=1.0, stride=3)` which will ensure
that when particles are added/removed/aligned etc. it will honor that
this property has 3 elements per particle.  The array itself is
flattened so in equations, users must remember to use the appropriate
striding.  This currently works on the CPU but not yet on the GPU or
with MPI.